requirements.txt
numpy>=1.18.0
nltk>=3.4
joblib
sklearn
sklearn-crfsuite
tensorflow>=2.1.0
transformers>=3.5.0
fastapi==0.61.2
uvicorn==0.11.3
discopy-data-rknaebel